+++ apps/web/src/store/models/FundingRound.js
@@ -12,6 +12,7 @@ FundingRound.modelName = 'FundingRound'
 
 FundingRound.fields = {
   id: attr(),
+  allowLateJoiners: attr(),
   allowSelfVoting: attr(),
   canSubmit: attr(),
   canVote: attr(),